# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Vanessa's Employee API doc",
        "version": "0.1.0",
        "description":
          "This is a simple CRUD API application made with Express and documented with Swagger",
        "contact":{
          "name": "Vanessa Hirwa G",
          "url": "vava.com",
          "email":"vanessahirwa5@gmail.com"
        }
      },
    "host": "localhost:1500",
    "basePath": "/",
    "tags": [
        {
            "name": "User",
            "description": "User API"
        },
        {
            "name": "Employee",
            "description": "Employee API"
        }
    ],
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "scheme": "bearer",
            "in": "header"
        }
    },
    "paths": {
        "/users/signup": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "description": "Create User",
                "operationId": "createUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {   
                        "in": "body",
                        "name": "body",
                        "description": "User object that needs to be added",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserSignUp"
                        }
                    }
                }
            }
        },
        "/users/login":{
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login User",
                "description": "Login User",
                "operationId": "userLogin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"  
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object that needs to be added",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                }
            }
        },
        "/employees/new" : {
            "post" : {
                "tags": [
                    "Employee"
                ],
                "summary": "Create Employee",
                "description": "Create employee",
                "operationId": "createEmployee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [{
                    "bearerAuth": "[]"
                }], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Employee object that needs to be added",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            }
        },
        "/employees" : {
            "get" : {
                "tags": [
                    "Employee"
                ],
                "summary": "Get all employees",
                "description": "Get all employees",
                "operationId": "getAllEmployees",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [{
                    "bearerAuth": "[]"
                }], 
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            }
        },
        "/employees/update/:id":{
            "put": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update Employee",
                "description": "Update employee",
                "operationId": "updateEmployee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [{
                    "bearerAuth": "[]"
                }], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Employee object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            }

        },
        "/employees/delete/:id":{
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete Employee",
                "description": "Delete employee",
                "operationId": "deleteEmployee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [{
                    "bearerAuth": "[]"
                }], 
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            }

        },
        "/employees/:id":{
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get Employee",
                "description": "Get employee",
                "operationId": "getEmployee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "security": [{
                    "bearerAuth": "[]"
                }], 
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Employee"
                        }
                    }
                }
            }
        
        }

    
    },
    "definitions": {
        "UserSignUp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "UserLogin": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "Employee" : {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "national_id": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "laptop_manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "number"
                }
            }
        }
    }
}